Making requested refactoring
