+++ src/main/java/uk/gov/crowncommercial/dts/scale/cat/repo/GCloudAssessmentResultRepo.java
@@ -1,6 +1,7 @@
 package uk.gov.crowncommercial.dts.scale.cat.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import uk.gov.crowncommercial.dts.scale.cat.model.entity.ca.GCloudAssessmentResult;
 
 import java.util.Set;
